feat: resolve modules module
